pyproject.toml
[project]
name = "osiris-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.11.0.post1",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.1",
    "python-lsp-server[all]>=1.13.0",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.0.0",
    "tiktoken>=0.9.0",
    "tree-sitter==0.20.2",
    "tree-sitter-languages==1.10.2",
    "tree-sitter-python>=0.23.6",
    "unidiff>=0.7.5",
]
